+++ nomenclator/src/main/java/flowmanager/nomenclator/service/TeamService.java
@@ -23,6 +23,12 @@ public class TeamService {
     private final OrganizationRepository organizationRepository;
     private final TeamMapper teamMapper;
 
+    private Team getTeam(Integer teamId) {
+        return teamRepository.findById(teamId).orElseThrow(
+                () -> new NotFoundException(String.format("Team with id %d not found", teamId))
+        );
+    }
+
     public List<TeamSummaryDto> findAllTeams() {
         return teamRepository
                 .findAll()
@@ -37,18 +43,30 @@ public class TeamService {
         ));
     }
 
+    @Transactional
+    protected List<User> getMembers(List<Integer> membersIds) {
+        List<User> users = userRepository.findAllById(membersIds);
+        if(users.size() != membersIds.size()) {
+            throw new NotFoundException("One or more users were not found");
+        }
+        return users;
+    }
+
+    @Transactional
     public TeamResponseDto createTeam(TeamCreateDto teamCreateDto) {
         Organization organization = organizationRepository.findById(teamCreateDto.getOrganizationId()).orElseThrow(
                 () -> new NotFoundException(String.format("Organization with id %d not found", teamCreateDto.getOrganizationId()))
         );
         Team team = teamMapper.toEntity(teamCreateDto, organization);
+        if (teamCreateDto.getMembersIds() != null && !teamCreateDto.getMembersIds().isEmpty()) {
+            List<User> members = getMembers(teamCreateDto.getMembersIds());
+            team.setMembers(members);
+        }
         return teamMapper.toResponseDto(teamRepository.save(team));
     }
 
     public TeamResponseDto updateTeam(Integer teamId, TeamUpdateDto teamUpdateDto) {
-        Team team = teamRepository.findById(teamId).orElseThrow(
-                () -> new NotFoundException(String.format("Team with id %d not found", teamId))
-        );
+        Team team = getTeam(teamId);
         Organization organization = team.getOrganization();
         if(teamUpdateDto.getOrganizationId() != null) {
             organization = organizationRepository.findById(teamUpdateDto.getOrganizationId()).orElseThrow(
@@ -68,17 +86,11 @@ public class TeamService {
 
     @Transactional
     public TeamResponseDto assignUsers(Integer teamId, TeamAssignDto teamAssignDto) {
-        Team team = teamRepository.findById(teamId).orElseThrow(
-                () -> new NotFoundException(String.format("Team with id %d not found", teamId))
-        );
+        Team team = getTeam(teamId);
 
-        List<User> users = userRepository.findAllById(teamAssignDto.getAssigneesIds());
-        if(users.size() != teamAssignDto.getAssigneesIds().size()) {
-            throw new NotFoundException("One or more users were not found");
-        }
-
-        team.setUsers(users);
-        users.forEach(user -> {
+        List<User> members = getMembers(teamAssignDto.getMembersIds());
+        team.setMembers(members);
+        members.forEach(user -> {
             if(!user.getAssignedTeams().contains(team)) {
                 user.getAssignedTeams().add(team);
             }
@@ -88,6 +100,13 @@ public class TeamService {
     }
 
     public void deleteTeam(Integer teamId) {
+        Team team = teamRepository.findById(teamId).orElse(null);
+        if(team == null) {
+            return;
+        }
+
+        team.getMembers().clear();
+        team.getProjects().clear();
         teamRepository.deleteById(teamId);
     }
 }
